Update runCode.js
Added some small changes. Instead of checking if the file includes .en check if it endsWith .en to prevent files like index.en.js or something. Also to add to the uniqueness I made the comments be !! instead of // and you can remove the new comment feature if you dont think its a good idea